requirements.txt
Django
gunicorn
psycopg2
PyYAML
djangorestframework
docker-py
dockerpty
docopt
requests
django-cors-headers
